pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.6.1"
description = ""
authors = ["bugaga427 <vladislav3372@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
isort = "^5.7.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
